Use traces to optimize interpreter performance

+++ src/instructions/common.rs
@@ -15,8 +15,8 @@ pub fn add<Mac: Machine>(
     rs1: RegisterIndex,
     rs2: RegisterIndex,
 ) {
-    let rs1_value = &machine.registers()[rs1];
-    let rs2_value = &machine.registers()[rs2];
+    let rs1_value = &machine.registers()[rs1 as usize];
+    let rs2_value = &machine.registers()[rs2 as usize];
     let value = rs1_value.overflowing_add(&rs2_value);
     update_register(machine, rd, value);
 }
@@ -27,8 +27,8 @@ pub fn addw<Mac: Machine>(
     rs1: RegisterIndex,
     rs2: RegisterIndex,
 ) {
-    let rs1_value = &machine.registers()[rs1];
-    let rs2_value = &machine.registers()[rs2];
+    let rs1_value = &machine.registers()[rs1 as usize];
+    let rs2_value = &machine.registers()[rs2 as usize];
     let value = rs1_value.overflowing_add(&rs2_value);
     update_register(machine, rd, value.sign_extend(&Mac::REG::from_usize(32)));
 }
@@ -39,8 +39,8 @@ pub fn sub<Mac: Machine>(
     rs1: RegisterIndex,
     rs2: RegisterIndex,
 ) {
-    let rs1_value = &machine.registers()[rs1];
-    let rs2_value = &machine.registers()[rs2];
+    let rs1_value = &machine.registers()[rs1 as usize];
+    let rs2_value = &machine.registers()[rs2 as usize];
     let value = rs1_value.overflowing_sub(&rs2_value);
     update_register(machine, rd, value);
 }
@@ -51,8 +51,8 @@ pub fn subw<Mac: Machine>(
     rs1: RegisterIndex,
     rs2: RegisterIndex,
 ) {
-    let rs1_value = &machine.registers()[rs1];
-    let rs2_value = &machine.registers()[rs2];
+    let rs1_value = &machine.registers()[rs1 as usize];
+    let rs2_value = &machine.registers()[rs2 as usize];
     let value = rs1_value.overflowing_sub(&rs2_value);
     update_register(machine, rd, value.sign_extend(&Mac::REG::from_usize(32)));
 }
@@ -63,7 +63,7 @@ pub fn addi<Mac: Machine>(
     rs1: RegisterIndex,
     imm: Immediate,
 ) {
-    let value = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
+    let value = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
     update_register(machine, rd, value);
 }
 
@@ -73,7 +73,7 @@ pub fn addiw<Mac: Machine>(
     rs1: RegisterIndex,
     imm: Immediate,
 ) {
-    let value = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
+    let value = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
     update_register(machine, rd, value.sign_extend(&Mac::REG::from_usize(32)));
 }
 
@@ -86,7 +86,7 @@ pub fn lb<Mac: Machine>(
     rs1: RegisterIndex,
     imm: Immediate,
 ) -> Result<(), Error> {
-    let address = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
+    let address = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
     let value = machine.memory_mut().load8(&address)?;
     // sign-extened
     update_register(machine, rd, value.sign_extend(&Mac::REG::from_usize(8)));
@@ -99,7 +99,7 @@ pub fn lh<Mac: Machine>(
     rs1: RegisterIndex,
     imm: Immediate,
 ) -> Result<(), Error> {
-    let address = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
+    let address = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
     let value = machine.memory_mut().load16(&address)?;
     // sign-extened
     update_register(machine, rd, value.sign_extend(&Mac::REG::from_usize(16)));
@@ -112,7 +112,7 @@ pub fn lw<Mac: Machine>(
     rs1: RegisterIndex,
     imm: Immediate,
 ) -> Result<(), Error> {
-    let address = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
+    let address = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
     let value = machine.memory_mut().load32(&address)?;
     update_register(machine, rd, value.sign_extend(&Mac::REG::from_usize(32)));
     Ok(())
@@ -124,7 +124,7 @@ pub fn ld<Mac: Machine>(
     rs1: RegisterIndex,
     imm: Immediate,
 ) -> Result<(), Error> {
-    let address = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
+    let address = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
     let value = machine.memory_mut().load64(&address)?;
     update_register(machine, rd, value.sign_extend(&Mac::REG::from_usize(64)));
     Ok(())
@@ -136,7 +136,7 @@ pub fn lbu<Mac: Machine>(
     rs1: RegisterIndex,
     imm: Immediate,
 ) -> Result<(), Error> {
-    let address = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
+    let address = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
     let value = machine.memory_mut().load8(&address)?;
     update_register(machine, rd, value);
     Ok(())
@@ -148,7 +148,7 @@ pub fn lhu<Mac: Machine>(
     rs1: RegisterIndex,
     imm: Immediate,
 ) -> Result<(), Error> {
-    let address = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
+    let address = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
     let value = machine.memory_mut().load16(&address)?;
     update_register(machine, rd, value);
     Ok(())
@@ -160,7 +160,7 @@ pub fn lwu<Mac: Machine>(
     rs1: RegisterIndex,
     imm: Immediate,
 ) -> Result<(), Error> {
-    let address = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
+    let address = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
     let value = machine.memory_mut().load32(&address)?;
     update_register(machine, rd, value);
     Ok(())
@@ -175,8 +175,8 @@ pub fn sb<Mac: Machine>(
     rs2: RegisterIndex,
     imm: Immediate,
 ) -> Result<(), Error> {
-    let address = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
-    let value = machine.registers()[rs2].clone();
+    let address = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
+    let value = machine.registers()[rs2 as usize].clone();
     machine.memory_mut().store8(&address, &value)?;
     Ok(())
 }
@@ -187,8 +187,8 @@ pub fn sh<Mac: Machine>(
     rs2: RegisterIndex,
     imm: Immediate,
 ) -> Result<(), Error> {
-    let address = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
-    let value = machine.registers()[rs2].clone();
+    let address = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
+    let value = machine.registers()[rs2 as usize].clone();
     machine.memory_mut().store16(&address, &value)?;
     Ok(())
 }
@@ -199,8 +199,8 @@ pub fn sw<Mac: Machine>(
     rs2: RegisterIndex,
     imm: Immediate,
 ) -> Result<(), Error> {
-    let address = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
-    let value = machine.registers()[rs2].clone();
+    let address = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
+    let value = machine.registers()[rs2 as usize].clone();
     machine.memory_mut().store32(&address, &value)?;
     Ok(())
 }
@@ -211,8 +211,8 @@ pub fn sd<Mac: Machine>(
     rs2: RegisterIndex,
     imm: Immediate,
 ) -> Result<(), Error> {
-    let address = machine.registers()[rs1].overflowing_add(&Mac::REG::from_i32(imm));
-    let value = machine.registers()[rs2].clone();
+    let address = machine.registers()[rs1 as usize].overflowing_add(&Mac::REG::from_i32(imm));
+    let value = machine.registers()[rs2 as usize].clone();
     machine.memory_mut().store64(&address, &value)?;
     Ok(())
 }
@@ -226,8 +226,8 @@ pub fn and<Mac: Machine>(
     rs1: RegisterIndex,
     rs2: RegisterIndex,
 ) {
-    let rs1_value = machine.registers()[rs1].clone();
-    let rs2_value = machine.registers()[rs2].clone();
+    let rs1_value = machine.registers()[rs1 as usize].clone();
+    let rs2_value = machine.registers()[rs2 as usize].clone();
     let value = rs1_value & rs2_value;
     update_register(machine, rd, value);
 }
@@ -238,8 +238,8 @@ pub fn xor<Mac: Machine>(
     rs1: RegisterIndex,
     rs2: RegisterIndex,
 ) {
-    let rs1_value = machine.registers()[rs1].clone();
-    let rs2_value = machine.registers()[rs2].clone();
+    let rs1_value = machine.registers()[rs1 as usize].clone();
+    let rs2_value = machine.registers()[rs2 as usize].clone();
     let value = rs1_value ^ rs2_value;
     update_register(machine, rd, value);
 }
@@ -250,8 +250,8 @@ pub fn or<Mac: Machine>(
     rs1: RegisterIndex,
     rs2: RegisterIndex,
 ) {
-    let rs1_value = machine.registers()[rs1].clone();
-    let rs2_value = machine.registers()[rs2].clone();
+    let rs1_value = machine.registers()[rs1 as usize].clone();
+    let rs2_value = machine.registers()[rs2 as usize].clone();
     let value = rs1_value | rs2_value;
     update_register(machine, rd, value);
 }
@@ -262,7 +262,7 @@ pub fn andi<Mac: Machine>(
     rs1: RegisterIndex,
     imm: Immediate,
 ) {
-    let value = machine.registers()[rs1].clone() & Mac::REG::from_i32(imm);
+    let value = machine.registers()[rs1 as usize].clone() & Mac::REG::from_i32(imm);
     update_register(machine, rd, value);
 }
 
@@ -272,12 +272,12 @@ pub fn xori<Mac: Machine>(
     rs1: RegisterIndex,
     imm: Immediate,
 ) {
-    let value = machine.registers()[rs1].clone() ^ Mac::REG::from_i32(imm);
+    let value = machine.registers()[rs1 as usize].clone() ^ Mac::REG::from_i32(imm);
     update_register(machine, rd, value);
 }
 
 pub fn ori<Mac: Machine>(machine: &mut Mac, rd: RegisterIndex, rs1: RegisterIndex, imm: Immediate) {
-    let value = machine.registers()[rs1].clone() | Mac::REG::from_i32(imm);
+    let value = machine.registers()[rs1 as usize].clone() | Mac::REG::from_i32(imm);
     update_register(machine, rd, value);
 }
 
@@ -287,7 +287,7 @@ pub fn slli<Mac: Machine>(
     rs1: RegisterIndex,
     shamt: UImmediate,
 ) {
-    let value = machine.registers()[rs1].clone() << Mac::REG::from_u32(shamt);
+    let value = machine.registers()[rs1 as usize].clone() << Mac::REG::from_u32(shamt);
     update_register(machine, rd, value);
 }
 
@@ -297,7 +297,7 @@ pub fn srli<Mac: Machine>(
     rs1: RegisterIndex,
     shamt: UImmediate,
 ) {
-    let value = machine.registers()[rs1].clone() >> Mac::REG::from_u32(shamt);
+    let value = machine.registers()[rs1 as usize].clone() >> Mac::REG::from_u32(shamt);
     update_register(machine, rd, value);
 }
 
@@ -307,7 +307,7 @@ pub fn srai<Mac: Machine>(
     rs1: RegisterIndex,
     shamt: UImmediate,
 ) {
-    let value = machine.registers()[rs1].signed_shr(&Mac::REG::from_u32(shamt));
+    let value = machine.registers()[rs1 as usize].signed_shr(&Mac::REG::from_u32(shamt));
     update_register(machine, rd, value);
 }
 
@@ -317,7 +317,7 @@ pub fn slliw<Mac: Machine>(
     rs1: RegisterIndex,
     shamt: UImmediate,
 ) {
-    let value = machine.registers()[rs1].clone() << Mac::REG::from_u32(shamt);
+    let value = machine.registers()[rs1 as usize].clone() << Mac::REG::from_u32(shamt);
     update_register(machine, rd, value.sign_extend(&Mac::REG::from_usize(32)));
 }
 
@@ -327,7 +327,7 @@ pub fn srliw<Mac: Machine>(
     rs1: RegisterIndex,
     shamt: UImmediate,
 ) {
-    let value = machine.registers()[rs1].zero_extend(&Mac::REG::from_usize(32))
+    let value = machine.registers()[rs1 as usize].zero_extend(&Mac::REG::from_usize(32))
         >> Mac::REG::from_u32(shamt);
     update_register(machine, rd, value.sign_extend(&Mac::REG::from_usize(32)));
 }
@@ -338,7 +338,7 @@ pub fn sraiw<Mac: Machine>(
     rs1: RegisterIndex,
     shamt: UImmediate,
 ) {
-    let value = machine.registers()[rs1]
+    let value = machine.registers()[rs1 as usize]
         .sign_extend(&Mac::REG::from_usize(32))
         .signed_shr(&Mac::REG::from_u32(shamt));
     update_register(machine, rd, value.sign_extend(&Mac::REG::from_usize(32)));
